# consumerService/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Example consumer API",
        "title": "Go PubSub Demo Consumer API",
        "contact": {
            "url": "https://github.com/samuel-poirier/go-pubsub-demo"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/samuel-poirier/go-pubsub-demo/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/api/v1/hc": {
            "get": {
                "description": "Returns ok if healthy",
                "produces": [
                    "application/json"
                ],
                "summary": "Health check endpoint",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
        "/api/v1/items/processed": {
            "get": {
                "description": "Returns top 100 last processed items",
                "produces": [
                    "application/json"
                ],
                "summary": "Endpoint to get processed items",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Number of items to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of items to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_samuel-poirier_go-pubsub-demo_consumer_internal_repository.ProcessedItem"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/items/processed/count": {
            "get": {
                "description": "Returns count of all processed items",
                "produces": [
                    "application/json"
                ],
                "summary": "Endpoint to count processed items",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "number"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_samuel-poirier_go-pubsub-demo_consumer_internal_repository.ProcessedItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "processedAt": {
                    "type": "string"
                },
                "processedData": {
                    "type": "string"
                }
            }
        }
    }
}