# consumerService/internal/repository/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package repository

import (
	"time"

	"github.com/google/uuid"
)

type ProcessedItem struct {
	ID            uuid.UUID
	ProcessedAt   time.Time
	ProcessedData string
}